requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-heroku==0.3.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
idna==2.8
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
